+++ ryders-rally/src/components/Events.jsx
@@ -3,7 +3,7 @@ import { FETCH_EVENTS_QUERY } from '../helper/graphql';
 import { AuthContext } from '../context/auth';
 import { useQuery } from '@apollo/react-hooks';
 import { Grid, Transition, Image } from 'semantic-ui-react';
-import "../styles/_event.scss"
+import "../_event.scss"
 
 
 function Events() {
