+++ ryders-rally/src/components/Nav.jsx
@@ -1,7 +1,7 @@
 import React, { useContext, useState } from 'react';
 import { Menu } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
-import "../styles/_nav.scss"
+import "../_nav.scss"
 
 import { AuthContext } from '../context/auth';
