+++ ryders-rally/src/components/_event.scss
