+++ ryders-rally/src/components/_home.scss
